+++ Source/ITIReact/index.ts
@@ -2,7 +2,7 @@
 export * from './DataUpdater'
 export * from './Inputs'
 export * from './Components'
+export * from './Routing'
 
 export * from './CancellablePromise'
-export * from './AsyncRouter'
 export * from './ITIReactContext'